move fetch_input to display module
This eliminates a silly dependency complication.
